unpack package
