+++ src/org/newdawn/spaceinvaders/ShopWindow.java
@@ -1,4 +1,4 @@
-package org.newdawn.spaceinvaders.windows;
+package org.newdawn.spaceinvaders;
 
 import java.awt.*;
 import java.awt.event.MouseAdapter;
