+++ src/org/newdawn/spaceinvaders/ShotEntity.java
@@ -1,6 +1,4 @@
-package org.newdawn.spaceinvaders.entity;
-
-import org.newdawn.spaceinvaders.cores.Game;
+package org.newdawn.spaceinvaders;
 
 /**
  * An entity representing a shot fired by the player's ship //총알
